+++ sz/src/sz_float_pwr.c
@@ -127,7 +127,7 @@ unsigned int optimize_intervals_float_1D_pwr(float *oriData, size_t dataLength, 
 	}
 	//compute the appropriate number
 	size_t targetCount = totalSampleSize*predThreshold;
-	int sum = 0;
+	size_t sum = 0;
 	for(i=0;i<maxRangeRadius;i++)
 	{
 		sum += intervals[i];
@@ -317,7 +317,7 @@ unsigned int optimize_intervals_float_2D_pwr(float *oriData, size_t r1, size_t r
 	}
 	//compute the appropriate number
 	size_t targetCount = totalSampleSize*predThreshold;
-	int sum = 0;
+	size_t sum = 0;
 	for(i=0;i<maxRangeRadius;i++)
 	{
 		sum += intervals[i];
@@ -488,7 +488,7 @@ unsigned int optimize_intervals_float_3D_pwr(float *oriData, size_t r1, size_t r
 	}
 	//compute the appropriate number
 	size_t targetCount = totalSampleSize*predThreshold;
-	int sum = 0;
+	size_t sum = 0;
 	for(i=0;i<maxRangeRadius;i++)
 	{
 		sum += intervals[i];
